Move s3 unit tests into correct package
Unit tests should be in the same package as the units they are testing
so that they can reach into internals without having to expose those
internals to the world.

Master issue: #1511
